Rename threading efficiency test limit to global variable fftw::testseconds.
Simplify integer division.

Update README.

+++ cconv.cc
@@ -111,7 +111,7 @@ int main(int argc, char* argv[])
   cout << "m=" << m << endl;
   
   if(N == 0) {
-    N=(int) ((double)N0/(double)n);
+    N=N0/n;
     if(N < 10) N=10;
   }
   cout << "N=" << N << endl;
